kselftests: Allow empty test name for run_many
An easy way to deal with skiplists for kselftest is to simply
remove the test from being executed in the main script. This
became a little more difficult with recent changes in the
framework as of Linux 5.2. (Those changes are very positive,
no doubt about it, just make skipping a bit harder.)

Using the following test suite as an example:
  run_many \
    "test_one.sh" \
    "test_two" \
    "test_three" \
    "test_four.sh"
if we were to skip the second test, the following would appear
in the runner:
  run_many \
    "test_one.sh" \
    "" \
    "test_three" \
    "test_four.sh"

The problem here is that the test counter is incremented, even
for a blank test. (Admittedly, it wasn't designed to work like
this.) This can be very easily overcomed by not taking an
empty test name into consideration.

Then, it can happen, like with this case:
  run_many \
    "lone_test.sh"
that removing that lone test will print the TAP13 header and
wrong test count, like this, which is altogether wrong:
  TAP version 13
  1..0

Accepting a blank test name will allow for cheap and easy
run-time skipping of tests in run_many.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>