uacce: Remove uacce mode 1 relatives
driver inclusion
category: feature
bugzilla: NA
CVE: NA

To be simple, in this patch  we try to remove uacce mode 1
related logic of uacce.c. Because in open mainline ,we do
not use this mode, mode 0 is used for kernel and mode 2 is
used for user.
At the same time, we update correspondingheader file uacce.h.
We also delete UACCE_QFRT_DKO in dummy_wd_dev.c and dummy_wd_v2.c

Signed-off-by: Yu'an Wang <wangyuan46@huawei.com>
Reviewed-by: Hui Tang <tanghui20@huawei.com>
Reviewed-by: Cheng Hu <hucheng.hu@huawei.com>
Signed-off-by: Yang Yingliang <yangyingliang@huawei.com>